requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn>=21.2.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.11.2
pydantic-settings==2.1.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
cryptography>=41.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pandas==2.1.3
numpy>=1.26.0,<2.0
pyngrok==7.0.0
email-validator>=2.0.0
psycopg2-binary>=2.9.0
